:recycle: 팟 삭제 예외처리 추가 (#172)

+++ src/main/java/com/tickettogether/domain/parts/service/MemberPartsServiceImpl.java
@@ -148,6 +148,11 @@ public class MemberPartsServiceImpl implements MemberPartsService {
         if (!parts.getManager().equals(user)) {
             throw new PartsDeleteDeniedException();
         }
+
+        if (parts.getCurrentPartTotal() > 1) {
+            throw new PartsMemberExistedException();
+        }
+
         partsRepository.deleteById(partId);
 
     }
